requirements.txt
Flask==2.0.1
grpcio==1.71.0
grpcio-tools==1.71.0
Pillow==11.2.1